'move to new framework and merge the truncate_conv and truncate_fc funcaiton'
